[Refactor] QueryRepository 형식에 맞추어 ChallengeParticipationRepository 구조 수정
