+++ keewe-domain/src/main/java/ccc/keewedomain/persistence/repository/challenge/ChallengeParticipationRepository.java
@@ -7,9 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ChallengeParticipationRepository extends
-        JpaRepository<ChallengeParticipation, Long>,
-        ChallengeParticipationQueryRepository {
+public interface ChallengeParticipationRepository extends JpaRepository<ChallengeParticipation, Long> {
 
     Optional<ChallengeParticipation> findByChallengerAndStatusAndDeletedFalse(User challenger, ChallengeParticipationStatus status);
 }
